Clean up and experiment built server
