+++ .github/workflows/pull_request.yml
@@ -16,3 +16,7 @@ jobs:
         run: yarn lint
       - name: Tests
         run: yarn test
+      - name: Build
+        run: yarn build
+      - name: e2e
+        run: yarn test:e2e
