--- libs/websocket/jest.config.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/no-commonjs
-module.exports = require("@howdypix/jest-config");
