--- libs/websocket/lib/index.ts
@@ -1,4 +1,0 @@
-import * as client from "./client";
-import * as server from "./server";
-
-export { client, server };
