--- libs/websocket/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@howdypix/websocket",
-  "version": "0.1.0",
-  "license": "MIT",
-  "main": "./build/index.js",
-  "types": "./build/index.d.ts",
-  "files": [
-    "dist/*"
-  ],
-  "scripts": {
-    "dev": "tsc --watch",
-    "build": "rm -rf build && tsc",
-    "lint": "eslint lib/**/*.{ts,tsx}",
-    "typecheck": "tsc --noEmit",
-    "test": "echo TBD"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.5.5",
-    "@babel/core": "^7.5.5",
-    "@howdypix/babel-preset": "^0.1.0",
-    "@howdypix/eslint-config": "^0.1.0",
-    "@howdypix/jest-config": "^0.1.0",
-    "@howdypix/typescript-config": "^0.1.0",
-    "@types/express": "^4.17.1",
-    "@types/lodash": "^4.14.138",
-    "@types/websocket": "^0.0.40"
-  },
-  "dependencies": {
-    "@howdypix/utils": "^0.1.0",
-    "express": "^4.17.1",
-    "lodash": "^4.17.15",
-    "typescript": "^3.6.3",
-    "websocket": "^1.0.30"
-  }
-}
