--- libs/websocket/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@howdypix/typescript-config/base",
-  "exclude": ["node_modules", "**/*.test.ts"],
-  "include": ["lib/**/*.ts"],
-  "compilerOptions": {
-    "declaration": true,
-    "declarationDir": "./build",
-    "outDir": "./build"
-  }
-}
